+++ lib/adjust_config.dart
@@ -14,6 +14,7 @@ import 'package:adjust_sdk/adjust_session_success.dart';
 import 'package:flutter/services.dart';
 
 enum AdjustLogLevel { verbose, debug, info, warn, error, suppress }
+
 enum AdjustEnvironment { production, sandbox }
 
 typedef void AttributionCallback(AdjustAttribution attributionData);
@@ -50,6 +51,7 @@ class AdjustConfig {
   static const String AdRevenueSourceAdMostSource = 'admost_sdk';
   static const String AdRevenueSourceUnity = 'unity_sdk';
   static const String AdRevenueSourceHeliumChartboost = 'helium_chartboost_sdk';
+  static const String AdRevenueSourcePublisher = 'publisher_sdk';
 
   String _appToken;
   AdjustEnvironment _environment;
@@ -71,6 +73,8 @@ class AdjustConfig {
   bool? launchDeferredDeeplink;
   bool? needsCost;
   bool? preinstallTrackingEnabled;
+  bool? playStoreKidsAppEnabled;
+  bool? coppaCompliantEnabled;
   String? sdkPrefix;
   String? userAgent;
   String? defaultTracker;
@@ -212,6 +216,12 @@ class AdjustConfig {
     if (preinstallTrackingEnabled != null) {
       configMap['preinstallTrackingEnabled'] =
           preinstallTrackingEnabled.toString();
+    }
+    if (playStoreKidsAppEnabled != null) {
+      configMap['playStoreKidsAppEnabled'] = playStoreKidsAppEnabled.toString();
+    }
+    if (coppaCompliantEnabled != null) {
+      configMap['coppaCompliantEnabled'] = coppaCompliantEnabled.toString();
     }
     if (allowiAdInfoReading != null) {
       configMap['allowiAdInfoReading'] = allowiAdInfoReading.toString();