+++ CHANGELOG.md
@@ -1,3 +1,20 @@
+### Version 4.30.0 (9th June 2022)
+#### Added
+- Added ability to mark your app as COPPA compliant. You can enable this setting by setting `coppaCompliantEnabled` member of `AdjustConfig` instance to `true`.
+- Added ability to mark your Android app as app for the kids in accordance to Google Play Families policies. You can enable this setting by setting `playStoreKidsAppEnabled` member of `AdjustConfig` instance to `true`.
+- Added `checkForNewAttStatus` method to `Adjust` API to allow iOS apps to instruct to SDK to check if `att_status` might have changed in the meantime.
+- Added support for Generic ad revenue tracking.
+
+#### Changed
+- Reverted `compileSdkVersion` from `32` to `31` (https://github.com/adjust/flutter_sdk/pull/77).
+- Changed responses which is being returned when iOS specific API is being called on Android platform (https://github.com/adjust/flutter_sdk/issues/79 and https://github.com/adjust/flutter_sdk/issues/80).
+
+#### Native SDKs
+- [iOS@v4.30.0][ios_sdk_v4.30.0]
+- [Android@v4.30.1][android_sdk_v4.30.1]
+
+---
+
 ### Version 4.29.2 (18th February 2022)
 #### Added
 - Added support for `Unity` ad revenue tracking.
@@ -256,6 +273,7 @@
 [ios_sdk_v4.29.2]: https://github.com/adjust/ios_sdk/tree/v4.29.2
 [ios_sdk_v4.29.6]: https://github.com/adjust/ios_sdk/tree/v4.29.6
 [ios_sdk_v4.29.7]: https://github.com/adjust/ios_sdk/tree/v4.29.7
+[ios_sdk_v4.30.0]: https://github.com/adjust/ios_sdk/tree/v4.30.0
 
 [android_sdk_v4.17.0]: https://github.com/adjust/android_sdk/tree/v4.17.0
 [android_sdk_v4.18.0]: https://github.com/adjust/android_sdk/tree/v4.18.0
@@ -267,4 +285,4 @@
 [android_sdk_v4.27.0]: https://github.com/adjust/android_sdk/tree/v4.27.0
 [android_sdk_v4.28.1]: https://github.com/adjust/android_sdk/tree/v4.28.1
 [android_sdk_v4.28.5]: https://github.com/adjust/android_sdk/tree/v4.28.5
-[android_sdk_v4.29.1]: https://github.com/adjust/android_sdk/tree/v4.29.1
+[android_sdk_v4.30.1]: https://github.com/adjust/android_sdk/tree/v4.30.1